VideoCommon: move TCacheEntry to its own file, so it can be forward declared
